refactor: 디자인 시스템 파일 cere 패키지로 이동
