+++ app/src/main/java/com/coinkiri/coinkiri/ui/login/component/LoginBtn.kt
@@ -22,11 +22,11 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.coinkiri.coinkiri.R
-import com.coinkiri.coinkiri.ui.designsystem.theme.Black
-import com.coinkiri.coinkiri.ui.designsystem.theme.CoinkiriTheme
-import com.coinkiri.coinkiri.ui.designsystem.theme.Kakao
-import com.coinkiri.coinkiri.ui.designsystem.theme.Naver
-import com.coinkiri.coinkiri.ui.designsystem.theme.White
+import com.coinkiri.coinkiri.core.designsystem.theme.Black
+import com.coinkiri.coinkiri.core.designsystem.theme.CoinkiriTheme
+import com.coinkiri.coinkiri.core.designsystem.theme.Kakao
+import com.coinkiri.coinkiri.core.designsystem.theme.Naver
+import com.coinkiri.coinkiri.core.designsystem.theme.White
 
 @Composable
 fun LoginBtn(
