+++ app/src/main/java/com/coinkiri/coinkiri/ui/profile/ProfileScreen.kt
@@ -16,9 +16,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.coinkiri.coinkiri.R
-import com.coinkiri.coinkiri.ui.designsystem.component.topappbar.CoinkiriTopBar
-import com.coinkiri.coinkiri.ui.designsystem.theme.CoinkiriTheme
-import com.coinkiri.coinkiri.ui.designsystem.theme.White
+import com.coinkiri.coinkiri.core.designsystem.component.topappbar.CoinkiriTopBar
+import com.coinkiri.coinkiri.core.designsystem.theme.CoinkiriTheme
+import com.coinkiri.coinkiri.core.designsystem.theme.White
 import com.coinkiri.coinkiri.ui.profile.component.FirstSettingList
 import com.coinkiri.coinkiri.ui.profile.component.ProfileImg
 import com.coinkiri.coinkiri.ui.profile.component.SecondSettingList
