+++ app/src/main/java/com/coinkiri/coinkiri/ui/profile/component/ProfileImg.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.coinkiri.coinkiri.R
-import com.coinkiri.coinkiri.ui.designsystem.theme.White
+import com.coinkiri.coinkiri.core.designsystem.theme.White
 
 @Composable
 fun ProfileImg() {
