+++ app/src/main/java/com/coinkiri/coinkiri/ui/profile/component/SecondSettingList.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.coinkiri.coinkiri.ui.designsystem.theme.Black
+import com.coinkiri.coinkiri.core.designsystem.theme.Black
 
 @Composable
 fun SecondSettingList() {
