+++ app/src/main/java/com/coinkiri/coinkiri/ui/profile/component/ThirdSettingList.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.coinkiri.coinkiri.ui.designsystem.theme.Black
+import com.coinkiri.coinkiri.core.designsystem.theme.Black
 
 @Composable
 fun ThirdSettingList() {
