+++ app/src/main/java/com/coinkiri/coinkiri/ui/talkdetail/ChatScreen.kt
@@ -6,7 +6,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.coinkiri.coinkiri.ui.designsystem.theme.CoinkiriTheme
+import com.coinkiri.coinkiri.core.designsystem.theme.CoinkiriTheme
 import com.coinkiri.coinkiri.ui.talkdetail.component.ChatContent
 import com.coinkiri.coinkiri.ui.talkdetail.component.ChatScreenTopBar
