+++ app/src/main/java/com/coinkiri/coinkiri/ui/talkdetail/component/ChatInputField.kt
@@ -24,8 +24,8 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.coinkiri.coinkiri.ui.designsystem.theme.SemiBlue
-import com.coinkiri.coinkiri.ui.designsystem.theme.White
+import com.coinkiri.coinkiri.core.designsystem.theme.SemiBlue
+import com.coinkiri.coinkiri.core.designsystem.theme.White
 
 
 @Composable
