+++ app/src/main/java/com/coinkiri/coinkiri/ui/talkdetail/component/ChatScreenTopBar.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
-import com.coinkiri.coinkiri.ui.designsystem.theme.CoinkiriTheme
+import com.coinkiri.coinkiri.core.designsystem.theme.CoinkiriTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
